refactor: 패키지 위치 변경, public 접근지정자 삭제
